Move around components for CSSClass, FieldList, and Figure
