+++ src/components/ComponentFactory.js
@@ -22,8 +22,8 @@ import Describe from './Describe';
 import DriversIndexTiles from './DriversIndexTiles';
 import Emphasis from './Emphasis';
 import Extract from './Extract';
-import Field from './Field';
-import FieldList from './FieldList';
+import Field from './FieldList/Field';
+import FieldList from './FieldList/FieldList';
 import Figure from './Figure';
 import Footnote from './Footnote/Footnote';
 import FootnoteReference from './Footnote/FootnoteReference';
