+++ tests/unit/Lightbox.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import Lightbox from '../../src/components/Lightbox';
+import Lightbox from '../../src/components/Figure/Lightbox';
 // data for this component
 import mockData from './data/Figure.test.json';
